DiceGameV2
Added a startGame function, this way the user can choose if he wants to play again or not, simply by typing Y or N. However, even if I type N, it restarts, same goes for other input. Also added an if statement to check if the user entered a number between 2-12. The dice rolls are random every new game now too.